Move check_user_can_make_*_address to validators
